+++ package.json
@@ -49,7 +49,6 @@
     "@types/passport": "^1.0.4",
     "@types/passport-discord": "^0.1.3",
     "dotenv": "^8.2.0",
-    "nodemon": "^2.0.4",
-    "ts-node": "^9.0.0"
+    "nodemon": "^2.0.4"
   }
 }
