+++ server/index.js
@@ -4,6 +4,7 @@ const app = express();
 const port = process.env.PORT || 3000;
 const session = require("express-session");
 const passport = require("passport");
+const discordStrategy = require("./strategies/discord");
 
 const authRoute = require("./routes/auth");
 const dashboardRoute = require("./routes/dashboard");
@@ -18,7 +19,8 @@ app.use(
   })
 );
 
-app.use(passport);
+app.use(passport.initialize());
+app.use(passport.session());
 
 app.use("/auth", authRoute);
 app.use("/dashboard", dashboardRoute);
